+++ bootique-jdbc-hikaricp-instrumented/src/main/java/io/bootique/jdbc/instrumented/hikaricp/metrics/HikariMetricsBridge.java
@@ -8,7 +8,6 @@ import com.codahale.metrics.Timer;
 import com.zaxxer.hikari.metrics.IMetricsTracker;
 import com.zaxxer.hikari.metrics.PoolStats;
 import io.bootique.jdbc.instrumented.hikaricp.JdbcHikariCPInstrumentedModule;
-import io.bootique.metrics.MetricNaming;
 
 import java.util.concurrent.TimeUnit;
 
@@ -17,7 +16,6 @@ import java.util.concurrent.TimeUnit;
  */
 public class HikariMetricsBridge implements IMetricsTracker {
 
-    private static final MetricNaming METRIC_NAMING = MetricNaming.forModule(JdbcHikariCPInstrumentedModule.class);
 
     private final String dataSourceName;
     private final Timer connectionWaitTimer;
@@ -42,35 +40,35 @@ public class HikariMetricsBridge implements IMetricsTracker {
     }
 
     public static String connectionWaitMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "Wait");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "Wait");
     }
 
     public static String connectionUsageMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "Usage");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "Usage");
     }
 
     public static String connectionCreationMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "ConnectionCreation");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "ConnectionCreation");
     }
 
     public static String connectionTimeoutRateMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "ConnectionTimeoutRate");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "ConnectionTimeoutRate");
     }
 
     public static String activeConnectionsMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "ActiveConnections");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "ActiveConnections");
     }
 
     public static String totalConnectionsMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "TotalConnections");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "TotalConnections");
     }
 
     public static String idleConnectionsMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "IdleConnections");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "IdleConnections");
     }
 
     public static String pendingConnectionsMetric(String dataSourceName) {
-        return METRIC_NAMING.name("Pool", dataSourceName, "PendingConnections");
+        return JdbcHikariCPInstrumentedModule.METRIC_NAMING.name("Pool", dataSourceName, "PendingConnections");
     }
 
     @Override