+++ bootique-jdbc-tomcat-instrumented/src/main/java/io/bootique/jdbc/instrumented/tomcat/JdbcTomcatInstrumentedModule.java
@@ -8,9 +8,12 @@ import com.google.inject.Singleton;
 import io.bootique.jdbc.DataSourceFactory;
 import io.bootique.jdbc.JdbcModule;
 import io.bootique.jdbc.instrumented.tomcat.healthcheck.TomcatHealthChecks;
+import io.bootique.metrics.MetricNaming;
 import io.bootique.metrics.health.HealthCheckModule;
 
 public class JdbcTomcatInstrumentedModule implements Module {
+
+    public static final MetricNaming METRIC_NAMING = MetricNaming.forModule(JdbcTomcatInstrumentedModule.class);
 
     @Override
     public void configure(Binder binder) {